refactored Header: created FemmecubatorLogo, MobileHeader, DesktopHeader components
